refactor: imagecache to class
- 여러곳에서 사용되서 클래스로 만듬
