+++ iOS/IssueTracker/IssueTracker/Issue/IssueDetail/View/IssueDetailCellHeader.swift
@@ -12,7 +12,6 @@ class IssueDetailCellHeader: UIView {
     let profileImage = UIImageView(frame: CGRect(x: 20, y: 25, width: 40, height: 40))
     let author = UILabel(frame: CGRect(x: 70, y: 25, width: 120, height: 20))
     let time = UILabel(frame: CGRect(x: 70, y: 45, width: 120, height: 20))
-    let imageCache = NSCache<NSURL, UIImage>()
 
     override init(frame: CGRect) {
         super.init(frame: frame)
@@ -41,23 +40,10 @@ class IssueDetailCellHeader: UIView {
         
         time.text = "\(PastTime().agoTime(from: comment.createdAt))"
         guard let url = URL(string: comment.user.avatar ?? "") else { return }
-        downloadImage(url: url) { (avatarImage: UIImage) in
+        ImageCache.shared.downloadImage(url: url) { (avatarImage: UIImage) in
             DispatchQueue.main.async {
                 self.profileImage.image = avatarImage
             }
-        }
-    }
-    
-    func downloadImage(url: URL, callback: @escaping(_ itemImage: UIImage) -> Void) {
-        if let imageFromCache = imageCache.object(forKey: url as NSURL) { callback(imageFromCache); return }
-        var image: UIImage = #imageLiteral(resourceName: "Icon")
-        DispatchQueue.global().async {
-            guard let data = try? Data(contentsOf: url) else { callback(image); return }
-            if let downloadedImage = UIImage(data: data) {
-                image = downloadedImage
-                self.imageCache.setObject(image, forKey: url as NSURL)
-            }
-            callback(image)
         }
     }
 }
